requirements.txt
flask
numpy
Pillow
tensorflow-cpu==2.1.0
Werkzeug